[Refactor] View와 ViewModel 분리 및 Model 리팩토링
