+++ Bloom/Bloom/Models/Product.swift
@@ -14,3 +14,17 @@ struct Product: Codable, Identifiable {
     let imageName: String
     let status: String
 }
+
+struct ProductDetail: Codable, Identifiable {
+    let id: UUID
+    let name: String
+    let price: String
+    let imageName: String
+    let description: String
+    let packagingMethod: String
+    let storageMethod: String
+    let caution: String
+    let marketName: String
+    let marketStatus: String
+    let marketLocation: String
+}
