+++ Bloom/Bloom/ViewModels/MarketProductsViewModel.swift
@@ -1,0 +1,28 @@
+//
+//  MarketProductsViewModel.swift
+//  Bloom
+//
+//  Created by 조다은 on 6/10/24.
+//
+
+import Foundation
+
+class MarketProductsViewModel: ObservableObject {
+    @Published var products: [Product] = []
+    
+    init() {
+        loadProducts()
+    }
+    
+    func loadProducts() {
+        let dummyProducts = [
+            Product(id: UUID(), name: "고백공격 장미꽃", price: "19,900원", imageName: "flower1", status: "운영중"),
+            Product(id: UUID(), name: "세상에서 제일 예쁜 꽃", price: "15,000원", imageName: "flower2", status: "준비중"),
+            Product(id: UUID(), name: "고백공격 장미꽃", price: "19,900원", imageName: "flower1", status: "운영중"),
+            Product(id: UUID(), name: "세상에서 제일 예쁜 꽃", price: "15,000원", imageName: "flower2", status: "준비중"),
+            Product(id: UUID(), name: "로맨틱 튤립", price: "12,000원", imageName: "flower3", status: "운영중"),
+            Product(id: UUID(), name: "봄의 정원", price: "25,000원", imageName: "flower4", status: "운영중")
+        ]
+        self.products = dummyProducts
+    }
+}
