+++ Bloom/Bloom/ViewModels/ProductDetailViewModel.swift
@@ -1,0 +1,28 @@
+//
+//  ProductDetailViewModel.swift
+//  Bloom
+//
+//  Created by 조다은 on 6/10/24.
+//
+
+import Foundation
+
+class ProductDetailViewModel: ObservableObject {
+    @Published var product: ProductDetail
+    
+    init() {
+        self.product = ProductDetail(
+            id: UUID(),
+            name: "고백공격 장미꽃다발",
+            price: "19,900원",
+            imageName: "flower1",
+            description: "꽃에 대한 설명을 바바바박 적어요.\n이 부분에는 글자수 제한이 있으면 좋을 것 같습니다!\n다들 힘내서 프로젝트 잘 마무리하세요~~",
+            packagingMethod: "강남역 바로 옆 초 역세권 짱짱 꽃집이랍니다.이 부분에는 글자수 제한이 있으면 좋을 것 같습니다!",
+            storageMethod: "강남역 바로 옆 초 역세권 짱짱 꽃집이랍니다. 이 부분에는 글자수 제한이 있으면 좋을 것 같습니다!",
+            caution: "강남역 바로 옆 초 역세권 짱짱 꽃집이랍니다. 이 부분에는 글자수 제한이 있으면 좋을 것 같습니다!",
+            marketName: "가든 플라워 라우라",
+            marketStatus: "운영중",
+            marketLocation: "강남구 서초동"
+        )
+    }
+}
