+++ Bloom/Bloom/Views/ProductDetail/ProductDetailView.swift
@@ -7,40 +7,6 @@
 
 import SwiftUI
 
-struct ProductDetail: Identifiable {
-    let id: UUID
-    let name: String
-    let price: String
-    let imageName: String
-    let description: String
-    let packagingMethod: String
-    let storageMethod: String
-    let caution: String
-    let marketName: String
-    let marketStatus: String
-    let marketLocation: String
-}
-
-class ProductDetailViewModel: ObservableObject {
-    @Published var product: ProductDetail
-    
-    init() {
-        self.product = ProductDetail(
-            id: UUID(),
-            name: "고백공격 장미꽃다발",
-            price: "19,900원",
-            imageName: "flower1",
-            description: "꽃에 대한 설명을 바바바박 적어요.\n이 부분에는 글자수 제한이 있으면 좋을 것 같습니다!\n다들 힘내서 프로젝트 잘 마무리하세요~~",
-            packagingMethod: "강남역 바로 옆 초 역세권 짱짱 꽃집이랍니다.이 부분에는 글자수 제한이 있으면 좋을 것 같습니다!",
-            storageMethod: "강남역 바로 옆 초 역세권 짱짱 꽃집이랍니다. 이 부분에는 글자수 제한이 있으면 좋을 것 같습니다!",
-            caution: "강남역 바로 옆 초 역세권 짱짱 꽃집이랍니다. 이 부분에는 글자수 제한이 있으면 좋을 것 같습니다!",
-            marketName: "가든 플라워 라우라",
-            marketStatus: "운영중",
-            marketLocation: "강남구 서초동"
-        )
-    }
-}
-
 struct ProductDetailView: View {
     @StateObject private var viewModel = ProductDetailViewModel()
