+++ Bloom/Bloom/Views/SearchResults/SearchResultsView.swift
@@ -7,28 +7,6 @@
 
 import SwiftUI
 
-class SearchResultsViewModel: ObservableObject {
-    @Published var products: [Product] = []
-    @Published var searchText: String = ""
-    @Published var showOnlyOperating: Bool = false
-    
-    init() {
-        loadProducts()
-    }
-    
-    func loadProducts() {
-        let dummyProducts = [
-            Product(id: UUID(), name: "고백공격 장미꽃", price: "19,900원", imageName: "flower1", status: "운영중"),
-            Product(id: UUID(), name: "세상에서 제일 예쁜 꽃", price: "15,000원", imageName: "flower2", status: "준비중"),
-            Product(id: UUID(), name: "고백공격 장미꽃", price: "19,900원", imageName: "flower1", status: "운영중"),
-            Product(id: UUID(), name: "세상에서 제일 예쁜 꽃", price: "15,000원", imageName: "flower2", status: "준비중"),
-            Product(id: UUID(), name: "로맨틱 튤립", price: "12,000원", imageName: "flower3", status: "운영중"),
-            Product(id: UUID(), name: "봄의 정원", price: "25,000원", imageName: "flower4", status: "운영중")
-        ]
-        self.products = dummyProducts
-    }
-}
-
 struct SearchResultsView: View {
     @StateObject private var viewModel = SearchResultsViewModel()
     
@@ -136,7 +114,6 @@ struct SearchResultsView: View {
         .navigationBarHidden(true)
     }
 }
-
 
 #Preview {
     SearchResultsView()
